refactor(core): extract distribution update policy
